requirements.txt
django
django-pygments
gunicorn
django-heroku
requests
djangorestframework
djangorestframework-simplejwt